+++ source/owl/public/scene/component/components.h
@@ -14,6 +14,7 @@
 #include "Canvas.h"
 #include "CircleRenderer.h"
 #include "EntityLink.h"
+#include "FlyCamera.h"
 #include "Hierarchy.h"
 #include "ID.h"
 #include "LuaScript.h"
@@ -52,17 +53,18 @@ concept isComponent =
 		std::is_same_v<Component, AnimatedSpriteRenderer> || std::is_same_v<Component, BackgroundTexture> ||
 		std::is_same_v<Component, Camera> || std::is_same_v<Component, Canvas> ||
 		std::is_same_v<Component, CircleRenderer> || std::is_same_v<Component, EntityLink> ||
-		std::is_same_v<Component, Hierarchy> || std::is_same_v<Component, ID> || std::is_same_v<Component, LuaScript> ||
-		std::is_same_v<Component, NativeScript> || std::is_same_v<Component, PhysicBody> ||
-		std::is_same_v<Component, Player> || std::is_same_v<Component, PrefabLink> ||
-		std::is_same_v<Component, RaycastDoor> || std::is_same_v<Component, RaycastPushWall> ||
-		std::is_same_v<Component, RendererTag> || std::is_same_v<Component, SoundListener> ||
-		std::is_same_v<Component, SoundSource> || std::is_same_v<Component, SpriteRenderer> ||
-		std::is_same_v<Component, Tag> || std::is_same_v<Component, Text> || std::is_same_v<Component, Tilemap> ||
-		std::is_same_v<Component, Transform> || std::is_same_v<Component, Trigger> ||
-		std::is_same_v<Component, UiButton> || std::is_same_v<Component, UiImage> ||
-		std::is_same_v<Component, UiPanel> || std::is_same_v<Component, UiProgressBar> ||
-		std::is_same_v<Component, UiRect> || std::is_same_v<Component, UiSlider> || std::is_same_v<Component, UiText> ||
+		std::is_same_v<Component, FlyCamera> || std::is_same_v<Component, Hierarchy> || std::is_same_v<Component, ID> ||
+		std::is_same_v<Component, LuaScript> || std::is_same_v<Component, NativeScript> ||
+		std::is_same_v<Component, PhysicBody> || std::is_same_v<Component, Player> ||
+		std::is_same_v<Component, PrefabLink> || std::is_same_v<Component, RaycastDoor> ||
+		std::is_same_v<Component, RaycastPushWall> || std::is_same_v<Component, RendererTag> ||
+		std::is_same_v<Component, SoundListener> || std::is_same_v<Component, SoundSource> ||
+		std::is_same_v<Component, SpriteRenderer> || std::is_same_v<Component, Tag> ||
+		std::is_same_v<Component, Text> || std::is_same_v<Component, Tilemap> || std::is_same_v<Component, Transform> ||
+		std::is_same_v<Component, Trigger> || std::is_same_v<Component, UiButton> ||
+		std::is_same_v<Component, UiImage> || std::is_same_v<Component, UiPanel> ||
+		std::is_same_v<Component, UiProgressBar> || std::is_same_v<Component, UiRect> ||
+		std::is_same_v<Component, UiSlider> || std::is_same_v<Component, UiText> ||
 		std::is_same_v<Component, Visibility> || std::is_same_v<Component, VoxelWorld>;
 
 /**
@@ -103,9 +105,9 @@ concept isDeserializableComponent =
  */
 using CopiableComponents =
 		std::tuple<Transform, Camera, Canvas, SpriteRenderer, AnimatedSpriteRenderer, CircleRenderer, Text, Tilemap,
-				   PhysicBody, Player, Trigger, EntityLink, BackgroundTexture, Visibility, Hierarchy, SoundSource,
-				   SoundListener, LuaScript, PrefabLink, RaycastDoor, RaycastPushWall, RendererTag, VoxelWorld,
-				   UiButton, UiImage, UiPanel, UiProgressBar, UiRect, UiSlider, UiText>;
+				   PhysicBody, Player, Trigger, EntityLink, FlyCamera, BackgroundTexture, Visibility, Hierarchy,
+				   SoundSource, SoundListener, LuaScript, PrefabLink, RaycastDoor, RaycastPushWall, RendererTag,
+				   VoxelWorld, UiButton, UiImage, UiPanel, UiProgressBar, UiRect, UiSlider, UiText>;
 
 /**
  * @brief
@@ -114,18 +116,19 @@ using CopiableComponents =
  */
 using SerializableComponents =
 		std::tuple<Tag, Transform, Camera, Canvas, SpriteRenderer, AnimatedSpriteRenderer, CircleRenderer, Text,
-				   Tilemap, PhysicBody, Player, Trigger, EntityLink, BackgroundTexture, Visibility, Hierarchy,
-				   SoundSource, SoundListener, LuaScript, PrefabLink, RaycastDoor, RaycastPushWall, RendererTag,
-				   VoxelWorld, UiButton, UiImage, UiPanel, UiProgressBar, UiRect, UiSlider, UiText>;
+				   Tilemap, PhysicBody, Player, Trigger, EntityLink, FlyCamera, BackgroundTexture, Visibility,
+				   Hierarchy, SoundSource, SoundListener, LuaScript, PrefabLink, RaycastDoor, RaycastPushWall,
+				   RendererTag, VoxelWorld, UiButton, UiImage, UiPanel, UiProgressBar, UiRect, UiSlider, UiText>;
 
 /**
  * @brief
  *  List all optional components.
  * @note All except ID, Tag & Transform that are mandatory.
  */
-using OptionalComponents = std::tuple<Camera, Canvas, SpriteRenderer, AnimatedSpriteRenderer, CircleRenderer, Text,
-									  Tilemap, PhysicBody, Player, Trigger, EntityLink, BackgroundTexture, SoundSource,
-									  SoundListener, LuaScript, PrefabLink, RaycastDoor, RaycastPushWall, RendererTag,
-									  VoxelWorld, UiButton, UiImage, UiPanel, UiProgressBar, UiRect, UiSlider, UiText>;
+using OptionalComponents =
+		std::tuple<Camera, Canvas, SpriteRenderer, AnimatedSpriteRenderer, CircleRenderer, Text, Tilemap, PhysicBody,
+				   Player, Trigger, EntityLink, FlyCamera, BackgroundTexture, SoundSource, SoundListener, LuaScript,
+				   PrefabLink, RaycastDoor, RaycastPushWall, RendererTag, VoxelWorld, UiButton, UiImage, UiPanel,
+				   UiProgressBar, UiRect, UiSlider, UiText>;
 
 }// namespace owl::scene::component